[#126] refactor(frontend): extract shared TabLoader and restructure risk level page as dynamic tab component
